// apps/forte-admin/src/components/Dashboard/Dashboard/types.ts
export type TickType = {
  label: string;
  value: number;
};

export type SegmentType = TickType & { color: string; className?: string };